+++ transport.go
@@ -1,0 +1,6 @@
+package jaeger
+
+type Transport interface {
+	Write([]byte) error
+	Close()
+}
